requirements.txt
PyYAML==6.0.3
colorama==0.4.6
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
numpy==2.2.6
ujson==5.11.0
wikipedia==1.4.0
Wikipedia-API==0.8.1
wolframalpha==5.1.3
transformers==4.57.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
requests==2.32.5
pydantic==2.12.4
psutil==7.1.3
pandas==2.3.3
bs4==0.0.2
aiohttp==3.13.2
numexpr==2.14.1
python-dotenv==1.2.1
astunparse==1.6.3